chore: remove dead code and unused abstractions
