+++ assets/js/components/Optimize/CopyButton.vue
@@ -12,6 +12,7 @@
 
 <script lang="ts">
 import { defineComponent } from "vue";
+import { copyWithFeedback } from "@/utils/clipboard";
 import "@h2d2/shopicons/es/regular/copy";
 import "@h2d2/shopicons/es/regular/checkmark";
 
@@ -52,54 +53,7 @@ export default defineComponent({
 	},
 	methods: {
 		handleCopy() {
-			// Modern browser API, requires secure context (localhost or https)
-			if (navigator.clipboard && window.isSecureContext) {
-				navigator.clipboard
-					.writeText(this.content)
-					.then(() => {
-						this.showCopiedFeedback();
-					})
-					.catch((err) => {
-						console.error("Failed to copy to clipboard:", err);
-						this.fallbackCopy();
-					});
-			} else {
-				// Fallback method for HTTP/non-secure contexts
-				this.fallbackCopy();
-			}
-		},
-		fallbackCopy() {
-			try {
-				// Create a temporary textarea element
-				const textarea = document.createElement("textarea");
-				textarea.value = this.content;
-				textarea.style.position = "fixed";
-				textarea.style.left = "-999999px";
-				textarea.style.top = "-999999px";
-				document.body.appendChild(textarea);
-
-				// Select and copy the text
-				textarea.focus();
-				textarea.select();
-				const successful = document.execCommand("copy");
-				document.body.removeChild(textarea);
-
-				if (successful) {
-					this.showCopiedFeedback();
-				} else {
-					alert("Copy failed. Please copy manually.");
-				}
-			} catch (err) {
-				console.error("Fallback copy failed:", err);
-				// Show user feedback that copy failed
-				alert("Copy failed. Please copy manually.");
-			}
-		},
-		showCopiedFeedback() {
-			this.copied = true;
-			setTimeout(() => {
-				this.copied = false;
-			}, 2000);
+			copyWithFeedback(this.content, (value) => (this.copied = value));
 		},
 	},
 });
